refactored out staticData from `$badapplerendition`
- refactored to use a standard modular subcommands structure
