--- commands/badapplerendition/README.md
@@ -1,3 +1,0 @@
-This command uses the following tables:
-
-- `data.Bad_Apple` [bad_apple.sql](bad_apple.sql)
