+++ commands/badapplerendition/index.js
@@ -1,4 +1,5 @@
-const { getLinkParser } = require("../../utils/link-parser.js");
+const subcommands = require("./subcommands");
+const subcommandNames = subcommands.map(i => i.name);
 
 module.exports = {
 	Name: "badapplerendition",
@@ -9,174 +10,41 @@ module.exports = {
 	Flags: ["mention","pipe"],
 	Params: null,
 	Whitelist_Response: null,
-	Static_Data: (() => ({
-		commands: [
-			{
-				name: "check",
-				description: "Checks if your provided link is in the database, and creates a suggestion to add, if it isn't."
-			},
-			{
-				name: "list",
-				description: "Simply posts the link to the Bad Apple!! list."
-			},
-			{
-				name: "random",
-				description: "Rolls a random rendition from the list, and posts its details."
-			}
-		]
-	})),
-	Code: (async function badAppleRendition (context, command, ...args) {
-		const commands = this.staticData.commands.map(i => i.name);
-		if (!command) {
+	Static_Data: null,
+	Code: async function badAppleRendition (context, commandName, ...args) {
+		if (!commandName) {
 			return {
 				success: false,
-				reply: `No command provided! Use one of these: ${commands.join(", ")}`
+				reply: `No command provided! Use one of these: ${subcommandNames.join(", ")}`
 			};
 		}
 
-		command = command.toLowerCase();
-		switch (command) {
-			case "check": {
-				const linkParser = getLinkParser();
-				const processed = new Set();
-				const results = [];
-
-				/** @type {string[]} */
-				const fixedArgs = args.flatMap(i => i.split(/\s+/).filter(Boolean));
-				for (const input of fixedArgs) {
-					const type = linkParser.autoRecognize(input);
-					if (type !== "youtube") {
-						continue;
-					}
-
-					const link = linkParser.parseLink(input);
-					if (processed.has(link)) {
-						continue;
-					}
-					else {
-						processed.add(link);
-					}
-
-					const existing = await sb.Query.getRecordset(rs => rs
-						.select("ID", "Device", "Link", "Timestamp")
-						.from("data", "Bad_Apple")
-						.where(`Link = %s OR JSON_SEARCH(Reuploads, "one", %s) IS NOT NULL`, link, link)
-						.limit(1)
-						.single()
-					);
-
-					if (existing) {
-						const timestamp = (existing.Timestamp) ? `?t=${existing.Timestamp}` : "";
-						results.push({
-							input,
-							reply: sb.Utils.tag.trim `
-								Link is in the list already:
-								https://supinic.com/data/bad-apple/detail/${existing.ID}
-								Bad Apple!! on ${existing.Device}
-								-
-								https://youtu.be/${existing.Link}${timestamp}
-							`
-						});
-						continue;
-					}
-
-					const data = await linkParser.fetchData(input);
-					if (!data) {
-						results.push({
-							input,
-							reply: `Video is not available!`
-						});
-
-						continue;
-					}
-
-					const row = await sb.Query.getRow("data", "Bad_Apple");
-					row.setValues({
-						Link: link,
-						Device: data.name,
-						Status: "Pending approval",
-						Type: null,
-						Published: data.created,
-						Notes: `Added to the list by ${context.user.Name}\n---\n${data.description ?? "No description"}`
-					});
-
-					const { insertId } = await row.save();
-					results.push({
-						input,
-						reply: `Link added to the rendition list, pending approval: https://supinic.com/data/bad-apple/detail/${insertId}`
-					});
-				}
-
-				if (results.length === 0) {
-					return {
-						success: false,
-						reply: `No proper links provided!`
-					};
-				}
-				else if (results.length === 1) {
-					return {
-						reply: results[0].reply
-					};
-				}
-				else {
-					const summary = results.map(i => `${i.input}\n${i.reply}`).join("\n\n");
-					const paste = await sb.Pastebin.post(summary);
-					if (paste.success !== true) {
-						return {
-							reply: `${results.length} videos processed. No summary available - ${paste.error ?? paste.body}`
-						};
-					}
-
-					return {
-						reply: `${results.length} videos processed. Summary: ${paste.body}`
-					};
-				}
-			}
-
-			case "list": {
-				return {
-					reply: `🍎 https://supinic.com/data/bad-apple/list`
-				};
-			}
-
-			case "random": {
-				const random = await sb.Query.getRecordset(rs => rs
-					.select("ID", "Device", "Link", "Timestamp")
-					.from("data", "Bad_Apple")
-					.where("Status = %s", "Approved")
-					.orderBy("RAND() DESC")
-					.limit(1)
-					.single()
-				);
-
-				const timestamp = (random.Timestamp) ? `?t=${random.Timestamp}` : "";
-				return {
-					reply: sb.Utils.tag.trim `
-						Bad Apple!! on ${random.Device}
-						https://youtu.be/${random.Link}${timestamp}
-						https://supinic.com/data/bad-apple/detail/${random.ID}
-					`
-				};
-			}
-
-			default: {
-				const commands = this.staticData.commands.map(i => i.name).join(", ");
-				return {
-					success: false,
-					reply: `Unrecognized command used! Pick one of these: ${commands}`
-				};
-			}
+		const subcommand = subcommands.find(i => i.name === commandName || i.aliases.includes(commandName));
+		if (!subcommand) {
+			return {
+				success: false,
+				reply: `Unknown command provided! Use one of these: ${subcommandNames.join(", ")}`
+			};
 		}
-	}),
-	Dynamic_Description: (async function (prefix) {
-		const staticData = this.staticData;
-		const subcommands = staticData.commands.map(i => `<li><code>${prefix}bar ${i.name}</code><br>${i.description}</li>`);
+
+		return await subcommand.execute(context, ...args);
+	},
+	Dynamic_Description: () => {
+		const subcommandList = subcommands.map(i => {
+			if (i.aliases.length === 0) {
+				return `<code>$bar ${i.name}</code>\n${i.description}`;
+			}
+			else {
+				const invocations = [i.name, ...i.aliases].map(invocation => `<code>$bar ${invocation}</code>`).join("\n");
+				return `${invocations}\n${i.description}`;
+			}
+		});
 
 		return [
 			"Aggregate command for all things Bad Apple!! related.",
 			"",
 
-			`<ul>${subcommands.join("")}</ul>`
+			`${subcommandList.join("")}`
 		];
-	})
+	}
 };
