+++ commands/badapplerendition/subcommands/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+	subcommands: [
+		require(`./check.js`),
+		require(`./list.js`),
+		require(`./random.js`)
+	]
+};
