+++ commands/badapplerendition/subcommands/list.js
@@ -1,0 +1,8 @@
+module.exports = {
+	name: "list",
+	aliases: ["show"],
+	description: "Simply posts the link to the Bad Apple!! list.",
+	execute: () => ({
+		reply: `🍎 https://supinic.com/data/bad-apple/list`
+	})
+};
